Added BackgroundService

+++ Application/Repositories/Base/IGenericRepository.cs
@@ -4,11 +4,11 @@ namespace Application.Repositories.Base;
 
 public interface IGenericRepository<T> where T : BaseEntity, new()
 {
-    Task<T?> GetAsync(int id);
+    Task<T?> GetAsync(long id);
     Task<IQueryable<T>> GetAllAsync();
 
 
     Task AddAsync(T entity);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
     Task<T> UpdateAsync(T entity);
 }
